pyproject.toml
[tool.poetry]
name = "metarepo"
version = "0.1.0"
description = "Git Meta Repository Manager"
authors = ["Jim Persson <jim@nurd.se>"]
license = "MIT"

[tool.poetry.scripts]
metarepo = "metarepo.cli:cli"
git-meta = "metarepo.cli:cli"

[tool.poetry.dependencies]
python = "^3.7"
gitpython = "^3.1.0"
cfgv = "^3.1.0"
pyyaml = "^5.3.1"
click = "^7.1.1"
colorama = "^0.4.3"

[tool.poetry.dev-dependencies]
tox = "^3.14.5"
pytest = "^5.3.5"
wheel = "^0.34.2"
black = "^19.10b0"
pre-commit = "^2.2.0"
pylint = "^2.4.4"
flake8 = "^3.7.9"

[tool.black]
line-length = 120

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
